Move fetching playbook to the parent component
